Move authentication to AuthGenerator
